refactor: redux + opa validation status icon
