+++ src/components/organisms/HotKeysHandler/HotKeysHandler.tsx
@@ -15,7 +15,6 @@ import {
   toggleRightMenu,
   toggleSettings,
   toggleStartProjectPane,
-  toggleValidationDrawer,
 } from '@redux/reducers/ui';
 import {
   currentConfigSelector,
@@ -244,7 +243,7 @@ const HotKeysHandler = () => {
   });
 
   useHotkeys(hotkeys.OPEN_VALIDATION_TAB.key, () => {
-    dispatch(toggleValidationDrawer());
+    dispatch(setLeftMenuSelection('validation'));
   });
 
   useHotkeys(hotkeys.RESET_RESOURCE_FILTERS.key, () => {
